feat: user group membership

+++ cmd/backend/internal/usergroup/handler.go
@@ -53,7 +53,17 @@ func (h *Handler) CreateUserGroup(ctx *gin.Context) {
 	})
 }
 
-func (h *Handler) DeleteUserGroupById(ctx *gin.Context) {
+func (h *Handler) AddUserToUserGroup(ctx *gin.Context) {
+	var addUserToUserGroupDTO usergroup.AddUserToUserGroupDTO
+
+	if err := ctx.ShouldBindJSON(&addUserToUserGroupDTO); err != nil {
+		ctx.JSON(http.StatusBadRequest, map[string]any{
+			"message": "invalid input",
+			"error":   err.Error(),
+		})
+		return
+	}
+
 	id := ctx.Param("id")
 	if len(id) == 0 {
 		ctx.JSON(http.StatusBadRequest, map[string]any{
@@ -62,7 +72,57 @@ func (h *Handler) DeleteUserGroupById(ctx *gin.Context) {
 		return
 	}
 
+	_, err := h.service.AddUserToUserGroup(ctx, addUserToUserGroupDTO, id)
+	if err != nil {
+		ctx.JSON(http.StatusInternalServerError, err)
+		return
+	}
+
+	ctx.JSON(http.StatusCreated, map[string]any{
+		"message":  "add user to group successfully",
+		"user_id":  addUserToUserGroupDTO.UserId,
+		"group_id": id,
+	})
+}
+
+func (h *Handler) DeleteUserGroupById(ctx *gin.Context) {
+	id := ctx.Param("group_id")
+	if len(id) == 0 {
+		ctx.JSON(http.StatusBadRequest, map[string]any{
+			"message": "empty id",
+		})
+		return
+	}
+
 	err := h.service.DeleteUserGroupById(ctx, id)
+	if err != nil {
+		ctx.JSON(http.StatusInternalServerError, err)
+		return
+	}
+
+	ctx.JSON(http.StatusOK, map[string]any{
+		"message": "deleted",
+	})
+}
+
+func (h *Handler) DeleteUserFromUserGroupById(ctx *gin.Context) {
+	gId := ctx.Param("group_id")
+	if len(gId) == 0 {
+		ctx.JSON(http.StatusBadRequest, map[string]any{
+			"message": "empty id",
+		})
+		return
+	}
+
+	uId := ctx.Param("user_id")
+	if len(uId) == 0 {
+		ctx.JSON(http.StatusBadRequest, map[string]any{
+			"message": "empty id",
+		})
+		return
+	}
+
+	err := h.service.DeleteUserFromUserGroupById(ctx, gId, uId)
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, err)
 		return
